+++ backend/.env.dist
@@ -1,3 +1,4 @@
 PORT=3001
 # create by: require('crypto').randomBytes(64).toString('hex')
 ACCESS_TOKEN_SECRET=
+ISSUER=
