+++ backend/neo4j.js
@@ -61,33 +61,3 @@ async function runQuery(response, query, parameter = {}) {
 }
 
 module.exports.runQuery = runQuery
-
-function getResult(res, filter, parameters, limit) {
-  const query = 'MATCH (result:Result)--(tag:Tag), (result)-[:GAME]->(game:Game), (player:Player)--(team:Team)-[score:SCORED]-(result) ' +
-    'WITH result, game, collect(ID(tag)) AS tagIds, collect(player.nick) AS names, collect(ID(player)) AS playerIds, collect(team.hash) AS team, collect(score.score) AS score, collect(score.funkies) AS funkies ' +
-    (filter.length > 0 ? 'WHERE ' + filter.join(' AND ') + ' ' : '') +
-    'RETURN result, game.name AS game, ID(game) AS gameID, names, playerIds, team, score, funkies ORDER BY result.date DESC' + limit;
-  return runQuery(res, query, parameters).then((results) => {
-    return results.map((result) => {
-      const scores = {}
-      result.team.forEach((team, index) => {
-        if (!scores[team]) {
-          scores[team] = {players: []}
-        }
-        scores[team].score = result.score[index]
-        scores[team].funkies = result.funkies[index]
-        scores[team].players.push({nick: result.names[index], id: result.playerIds[index]})
-      })
-      return {
-        ...result.result,
-        game: {
-          id: result.gameID,
-          name: result.game
-        },
-        scores: Object.values(scores)
-      }
-    })
-  })
-}
-
-module.exports.getResult = getResult
